Extract core actions
