+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -1,3 +1,4 @@
 export * from "./actions";
+export * from "./core";
 export * from "./ui";
 export * from "./writeback";
